refactor: split the canvas helpers out of the e2e spec
index.spec.ts held two concerns at 529 lines: nine canvas-measurement and
interaction helpers, and the scenario suite. The helpers move to a sibling
canvas.ts, following the pattern target.ts already set.

The move is byte-identical apart from the export keyword and one prettier
reflow. Reviewing it surfaced three things worth carrying:

- `> 24` and `/ 765` became LIT_THRESHOLD and MAX_PIXEL_SUM. Exporting the
  functions makes those constants public arithmetic; 3 * 255 is verified
  equal to 765 over 2000 random samples plus the empty and length-mismatch
  edges.
- meanBrightness and litFraction return -1 when the canvas cannot be
  sampled, which was undocumented. It is documented now, and the three
  callers that asserted only an upper bound — where -1 passes — assert a
  lower bound first.
- Two of those guards use >= 0, admitting every real reading and excluding
  only the sentinel. The third keeps > 0 because its own comment says a
  degenerate reading divides out to 1 and passes.

The decay guard was > 0 for one round; measured after=0.010 on this
machine, a couple of brightness units from zero, so a faster renderer would
have failed it. >= 0 is what the contract actually asks for.

Closes #719
